pyproject.toml
[tool.poetry]
name = "phone_reminder"
version = "0.1.0"
description = ""
authors = ["Lee-Roy King <leeroywking@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
twilio = "^7.3.0"
python-dotenv = "^0.19.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
